Created a mercurial repository for the 'array fill' projects group.
In November 2008 I created 3 projects with the intention to conduct
microbenchmarks that would show the difference in performance between native
C++, C++/CLI running inside a CLR console application and C++/CLI running in a
Windows Forms application when filling arrays of values. Later I conducted
similar microbenchmarks in variations of Java running on Sun's JRE.

This repository is intended to track the source code of these 3 projects each in
a different branch.